+++ src/StateSmith/Runner/Runner.cs
@@ -1,4 +1,4 @@
-﻿using StateSmith.output.C99BalancedCoder1;
+using StateSmith.output.C99BalancedCoder1;
 using StateSmith.output.UserConfig;
 using System;
 using System.Collections.Generic;
@@ -11,6 +11,7 @@ using StateSmith.Input.Expansions;
 using StateSmith.Input;
 using StateSmith.compiler.Visitors;
 using StateSmith.Input.PlantUML;
+using StateSmith.compiler;
 
 #nullable enable
 
@@ -63,23 +64,28 @@ namespace StateSmith.Runner
                     throw;
                 }
 
+                Environment.ExitCode = -1; // lets calling process know that code gen failed
+
                 exceptionPrinter.PrintException(e);
+                DumpErrorDetailsToFile(e);
                 OutputStageMessage("finished with failure.");
             }
 
             System.Console.WriteLine();
         }
 
+        private void DumpErrorDetailsToFile(Exception e)
+        {
+            var errorDetailFilePath = settings.diagramFile + ".err.txt";
+            //errorDetailFilePath = Path.GetFullPath(errorDetailFilePath); // if you want the full path but with resolved "../../"
+            errorDetailFilePath = Path.GetRelativePath(Directory.GetCurrentDirectory(), errorDetailFilePath);
+            exceptionPrinter.DumpExceptionDetails(e, errorDetailFilePath);
+            Console.Error.WriteLine("Additional exception detail dumped to file: " + errorDetailFilePath);
+        }
+
         protected void RunRest()
         {
-            if (settings.stateMachineName != null)
-            {
-                sm = (Statemachine)compiler.GetVertex(settings.stateMachineName);
-            }
-            else
-            {
-                sm = compiler.rootVertices.OfType<Statemachine>().Single();
-            }
+            FindStateMachine();
 
             CodeGenContext codeGenContext = new(sm, settings.renderConfig);
             settings.mangler.SetStateMachine(sm);
@@ -102,11 +108,30 @@ namespace StateSmith.Runner
             File.WriteAllText($"{settings.outputDirectory}{settings.mangler.CFileName}", cFileContents);
         }
 
+        private void FindStateMachine()
+        {
+            if (settings.stateMachineName != null)
+            {
+                var action = () => { sm = (Statemachine)compiler.GetVertex(settings.stateMachineName); };
+                action.RunOrWrapException((e) => new ArgumentException($"Couldn't find state machine in diagram with name `{settings.stateMachineName}`.", e));
+            }
+            else
+            {
+                var action = () => { sm = compiler.rootVertices.OfType<Statemachine>().Single(); };
+                action.RunOrWrapException((e) => new ArgumentException($"State machine name not specified. Expected diagram to have find 1 Statemachine node at root level. Instead, found {compiler.rootVertices.OfType<Statemachine>().Count()}.", e));
+            }
+
+            OutputStageMessage($"Generating code for state machine `{sm.Name}`.");
+        }
+
         private void OutputCompilingDiagramMessage()
         {
             string filePath = Path.GetRelativePath(AppDomain.CurrentDomain.BaseDirectory + "../../../..", settings.diagramFile);
             filePath = filePath.Replace('\\', '/');
-            OutputStageMessage($"Compiling file: `{filePath}`");
+            OutputStageMessage($"Compiling file: `{filePath}` "
+                + ((settings.stateMachineName == null) ? "(no state machine name specified)" : $"with target state machine name: `{settings.stateMachineName}`")
+                + "."
+            );
         }