+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [WIP]
+### Added
+- Runner - set process exit code to -1 on failure.
+- Runner - output additional exception details to `<diagram_file_path>.err.txt`.
+  This is useful for generic assert like exceptions that don't yet have useful error messages.
+  https://github.com/StateSmith/StateSmith/issues/38
+- validation - new exception: `State machines must have exactly 1 initial state. Actual count: 0.`
+- validation - helpful error messages for when state machine design is not found.
+
+---
+
 ## [0.5.7-alpha]
 ### Added
 - detect yEd hidden edges https://github.com/StateSmith/StateSmith/issues/29